Added ProvideSystemProperty rule.
The ProvideSystemProperty rule provides an arbitrary value for a system
property to a test. After the test the original value is restored. You
can ensure that a property is not set by providing null.

Let's assume the system property MyProperty is not set and the system
property OtherProperty has the value OtherValue. Now run the test

 public void MyTest {
   @Rule
   public final ProvideSystemProperty myPropertyHasMyValue
     = new ProvideSystemProperty("MyProperty", "MyValue");

   @Rule
   public final ProvideSystemProperty otherPropertyIsMissing
     = new ProvideSystemProperty("OtherProperty", null);

   @Test
   public void overrideProperty() {
     assertEquals("MyValue", System.getProperty("MyProperty"));
     assertNull(System.getProperty("OtherProperty"));
   }
 }

The test succeeds and after the test, the system property MyProperty is
not set and the system property OtherProperty has the value OtherValue.